feat: include resolver ids in IssueTransition event [36]
